Factor the note window gui
