+++ src/gui/mod.rs
@@ -1,1 +1,2 @@
 pub mod r#trait;
+pub mod notewindow;
